Move all `LocationProviders`-related code into locations.py
